Trait reth_network_api::NetworkInfo

pub trait NetworkInfo: Send + Sync {
    // Required methods
    fn local_addr(&self) -> SocketAddr;
    fn network_status<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<NetworkStatus, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn chain_id(&self) -> u64;
    fn is_syncing(&self) -> bool;
    fn is_initially_syncing(&self) -> bool;
}

Provides general purpose information about the network.

Required Methods

fn local_addr(&self) -> SocketAddr

Returns the SocketAddr that listens for incoming connections.

fn network_status<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<NetworkStatus, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the current status of the network being ran by the local node.

fn chain_id(&self) -> u64

Returns the chain id

fn is_syncing(&self) -> bool

Returns true if the network is undergoing sync.

fn is_initially_syncing(&self) -> bool

Returns true when the node is undergoing the very first Pipeline sync.