Trait reth_network_api::Peers

pub trait Peers: PeersInfo {

    // Required methods
    fn add_peer_kind(&self, peer: PeerId, kind: PeerKind, addr: SocketAddr);
    fn get_peers_by_kind<'life0, 'async_trait>(
        &'life0 self,
        kind: PeerKind
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_all_peers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_peer_by_id<'life0, 'async_trait>(
        &'life0 self,
        peer_id: PeerId
    ) -> Pin<Box<dyn Future<Output = Result<Option<PeerInfo>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_peers_by_id<'life0, 'async_trait>(
        &'life0 self,
        peer_ids: Vec<PeerId>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_peer(&self, peer: PeerId, kind: PeerKind);
    fn disconnect_peer(&self, peer: PeerId);
    fn disconnect_peer_with_reason(
        &self,
        peer: PeerId,
        reason: DisconnectReason
    );
    fn reputation_change(&self, peer_id: PeerId, kind: ReputationChangeKind);
    fn reputation_by_id<'life0, 'async_trait>(
        &'life0 self,
        peer_id: PeerId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Reputation>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn add_peer(&self, peer: PeerId, addr: SocketAddr) { ... }
    fn add_trusted_peer(&self, peer: PeerId, addr: SocketAddr) { ... }
    fn get_trusted_peers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_basic_peers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }

}

Provides an API for managing the peers of the network.

Required Methods

fn add_peer_kind(&self, peer: PeerId, kind: PeerKind, addr: SocketAddr)

Adds a peer to the known peer set, with the given kind.

fn get_peers_by_kind<'life0, 'async_trait>( &'life0 self, kind: PeerKind ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the rpc PeerInfo for all connected peers with the given kind.

fn get_all_peers<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the rpc PeerInfo for all connected peers.

fn get_peer_by_id<'life0, 'async_trait>( &'life0 self, peer_id: PeerId ) -> Pin<Box<dyn Future<Output = Result<Option<PeerInfo>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the rpc PeerInfo for the given peer id.

Returns None if the peer is not connected.

fn get_peers_by_id<'life0, 'async_trait>( &'life0 self, peer_ids: Vec<PeerId> ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the rpc PeerInfo for the given peers if they are connected.

Note: This only returns peers that are connected, unconnected peers are ignored but keeping the order in which they were requested.

fn remove_peer(&self, peer: PeerId, kind: PeerKind)

Removes a peer from the peer set that corresponds to given kind.

fn disconnect_peer(&self, peer: PeerId)

Disconnect an existing connection to the given peer.

fn disconnect_peer_with_reason(&self, peer: PeerId, reason: DisconnectReason)

Disconnect an existing connection to the given peer using the provided reason

fn reputation_change(&self, peer_id: PeerId, kind: ReputationChangeKind)

Send a reputation change for the given peer.

fn reputation_by_id<'life0, 'async_trait>( &'life0 self, peer_id: PeerId ) -> Pin<Box<dyn Future<Output = Result<Option<Reputation>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the reputation of a peer.

Provided Methods

fn add_peer(&self, peer: PeerId, addr: SocketAddr)

Adds a peer to the peer set.

fn add_trusted_peer(&self, peer: PeerId, addr: SocketAddr)

Adds a trusted peer to the peer set.

fn get_trusted_peers<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the rpc PeerInfo for all connected PeerKind::Trusted peers.

fn get_basic_peers<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<PeerInfo>, NetworkError>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the rpc PeerInfo for all connected PeerKind::Basic peers.