Trait reth_network_api::PeersInfo

pub trait PeersInfo: Send + Sync {
    // Required methods
    fn num_connected_peers(&self) -> usize;
    fn local_node_record(&self) -> NodeRecord;
}

Provides general purpose information about Peers in the network.

Required Methods

fn num_connected_peers(&self) -> usize

Returns how many peers the network is currently connected to.

Note: this should only include established connections and not ongoing attempts.

fn local_node_record(&self) -> NodeRecord

Returns the Ethereum Node Record of the node.